Fehlerbehandlung und Logging in BuchungsController
Es wurde eine Überprüfung hinzugefügt, ob die Liste `termin`
leer ist, und ein `NotFound()`-Ergebnis wird zurückgegeben,
falls keine Einträge vorhanden sind.

Ein `Console.WriteLine`-Befehl wurde hinzugefügt, um die
`KursTerminId` jedes Elements auszugeben.

Der Löschvorgang wurde um einen `try-catch`-Block erweitert,
um `DbUpdateException`-Fehler abzufangen. Im Fehlerfall wird
ein `BadRequest`-Ergebnis mit einer Fehlermeldung zurückgegeben.

+++ backend/Fitness/Fitness/Controllers/BuchungsController.cs
@@ -41,9 +41,23 @@ namespace Fitness.Controllers
                 .Where(u => u != null && u.User.Name == buchung.Name && u.User.Vorname == buchung.Vorname && u.KursTerminId == buchung.terminID)
                 .ToListAsync();
 
+            if (termin.Count() == 0)
+            {
+                return NotFound();
+            }
+
             foreach (var item in termin)
             {
+                Console.WriteLine(item.KursTerminId);
                 _context.NimmtTeil.Remove(item);
+
+                try
+                {
+                    await _context.SaveChangesAsync();
+                } catch(DbUpdateException ex)
+                {
+                    return BadRequest($"Error beim stonieren: {ex.Message}");
+                }
             }
 
             return NoContent();
